requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
google-genai>=1.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
